+++ crates/cli/src/dataset.rs
@@ -53,11 +53,7 @@ fn process(env: &Env, path: &Path, dataset: Dataset) -> Result<()> {
             .find(|p| p.extension() == Some(OsStr::new("fail"))),
     ) {
         (Some(good), Some(fail)) => {
-            let model = Model::train(
-                env,
-                [content_from_pathbuf(good.to_path_buf())].to_vec(),
-                logreduce_model::hashing_index::new,
-            )?;
+            let model = Model::train(env, [content_from_pathbuf(good.to_path_buf())].to_vec())?;
             let index = model.get_index(&IndexName::new()).unwrap();
             let anomalies = index
                 .inspect(